token on routes

+++ controller/comments-router.js
@@ -1,5 +1,6 @@
 const queries = require("../model/queries.js");
 const express = require("express");
+const jwt = require('jsonwebtoken');
 
 const router = express.Router();
 
@@ -8,8 +9,7 @@ router.get("/", (req, res) => {
   queries.readComments()
     .then((comments) => {
       res.status(200).json(comments);
-    })
-    .catch((error) => {
+    }).catch((error) => {
       res.status(500).json({ message: "Unable to perform the operation", error: error });
     });
 });
@@ -30,18 +30,21 @@ router.post("/", async (req, res) => {
 
 // DELETE comment
 router.delete("/", async (req, res) => {
-  const commentID = req.body.id;
-  await queries.deleteComment(commentID)
-    .then((count) => {
-      if (count > 0) { // Knex returns number of affected rows deleted.
-        res.status(200).json({ message: "Record successfully deleted." });
-      } else {
-        res.status(404).json({ message: "Record not found." });
-      }
-    })
-    .catch((error) => {
-      res.status(500).json({ message: "Unable to perform the operation.", error: error });
-    });
+
+  const token = req.headers['authorization'].split(' ')[1];
+  try {
+    jwt.verify(token, process.env.PRIVATE_KEY);
+
+    if (req.body.role === "admin") {
+      const commentID = req.body.id;
+      await queries.deleteComment(commentID)
+      res.status(200).json({ message: "Record successfully deleted" });
+    } else {
+      res.status(403).json({ message: "Client does not have access rights" }) 
+    }   
+  } catch {
+    res.status(401).json({ message: "Unauthorized. Client must authenticate to get the requested response." }) 
+  }
 });
 
 // PATCH comment votes
